requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
dill==0.3.4
Faker==10.0.0
idna==3.3
pyfunctional==1.4.3
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
tabulate==0.8.9
text-unidecode==1.3
urllib3==1.26.7
